sanitize function branch name and separate santize functions
