+++ backend/src/utils/utils.ts
@@ -13,18 +13,3 @@ const runShellCommand = async (command: string): Promise<string> => {
     throw new Error(error.message)
   }
 }
-
-export const sanitizeString = (inputStr: string):string => {
-  let str = inputStr.trim()
-  str = stripDangerousSpecialChars(str)
-  str = stripConsecutiveSpecialChars(str)
-  return str
-}
-
-const stripDangerousSpecialChars = (str: string):string => {
-  return str.replace(/[^a-zA-Z0-9\s+-.,?!]/g, '')
-}
-
-const stripConsecutiveSpecialChars = (str: string):string => {
-  return str.replace(/(\W)\1+/g, '$1')
-}
